requirements.txt
requests
python-dotenv
flask
flask-sqlalchemy
flask-login
psycopg2
